delet old, rename new text input
