+++ packages/ui/src/forms/TextInput.tsx
@@ -2,19 +2,25 @@
 
 import { IAttributes, getAttributes } from './attributesMap';
 
+export type TextInputClassNames = {
+  input?: string;
+  label?: string;
+  label_span?: string;
+  wrapper_div?: string;
+};
+
 type Props = {
   name: string;
   reg_name?: string;
   register: any;
   registeroptions?: any;
-  defaultValue?: string;
+  defaultValue?: string | number;
   label?: boolean;
-  labelClassName?: string;
   tooltip?: string;
-  className?: string;
+  className?: TextInputClassNames;
   maxW?: string;
   prepend?: string;
-  prependClassName?: string;
+  digits?: number;
 };
 
 /* Takes maxW bc unable to overwrite max-w with tailwind. This is weird as other styles can be overwritten */
@@ -26,17 +32,28 @@ export const TextInput = ({
   registeroptions,
   defaultValue,
   label = false,
-  labelClassName,
   tooltip,
   className,
   maxW = 'max-w-sm',
   prepend,
-  prependClassName,
+  digits = 3, // useed with type='number' to set width
 }: Props): JSX.Element => {
   const { _type, _label, _placeholder, _autocomplete, _rows, _maxLenght, _step }: IAttributes =
     getAttributes(name);
 
-  const register_string = reg_name ? reg_name : name;
+  const register_string = reg_name || name;
+  const isNumber = _type === 'number';
+
+  /* (digits + 2) * 4 */
+  const digitsToWidthClass: {
+    [key: number]: string;
+  } = {
+    1: 'w-12',
+    2: 'w-16',
+    3: 'w-20',
+    4: 'w-24',
+    5: 'w-28',
+  };
 
   let Input =
     _type === 'textarea' ? (
@@ -47,7 +64,7 @@ export const TextInput = ({
           placeholder={_placeholder}
           autoComplete={_autocomplete}
           defaultValue={defaultValue}
-          className={`textarea textarea-bordered w-full mx-auto focus:outline-none focus:border-accent ${maxW} ${className}`}
+          className={`textarea textarea-bordered w-full mx-auto focus:outline-none focus:border-accent ${maxW} ${className?.input}`}
           rows={_rows}
           maxLength={_maxLenght}
         />
@@ -55,28 +72,35 @@ export const TextInput = ({
     ) : (
       <>
         <input
-          // {...register(name, registeroptions)}
           {...register(register_string, registeroptions)}
           type={_type}
           placeholder={_placeholder}
           autoComplete={_autocomplete}
           defaultValue={defaultValue}
+          // max={isNumber ? '100' : undefined}
+          // maxlength='4'
+          // size={digits}
           step={_step}
           className={`${
-            _type ? 'flex-1 pl-0 h-auto focus:border-none' : `input-bordered ${maxW}`
-          } input w-full mx-auto focus:outline-none focus:border-accent ${className}`}
+            isNumber
+              ? `flex-1 text-center h-auto focus:border-none ${prepend ? 'px-0' : 'pr-0'} ${
+                  digitsToWidthClass[digits]
+                }`
+              : `input-bordered w-full ${maxW}`
+          } input mx-auto focus:outline-none focus:border-accent ${className?.input}`}
           maxLength={_maxLenght}></input>
       </>
     );
 
-  if (prepend) {
+  if (isNumber) {
     Input = (
-      // <div className='input input-bordered flex gap-2 w-full focus-within:[border:1px_solid_black] pl-0'>
       <div
-        className={`input input-bordered flex gap-2 w-full focus-within:[border:1px_solid_black] pl-0 ${maxW} ${prependClassName}`}>
-        <span className='text-sm bg-gray-200 h-full flex items-center justify-center px-2 my-auto'>
-          {prepend}
-        </span>
+        className={`input input-bordered flex focus-within:[border:1px_solid_black] px-0 ${maxW} ${className?.wrapper_div}`}>
+        {prepend && (
+          <span className='text-sm bg-gray-200 h-full flex items-center justify-center px-2 my-auto'>
+            {prepend}
+          </span>
+        )}
         {Input}
       </div>
     );
@@ -93,9 +117,9 @@ export const TextInput = ({
   if (label) {
     Input = (
       <>
-        {/* <label className='label w-full flex flex-col'> */}
-        <label className={`label w-full flex flex-col ${labelClassName}`}>
-          <span className='label-text self-start mb-3'>{_label}</span>
+        <label
+          className={`label w-full text-center flex flex-col whitespace-nowrap ${className?.label}`}>
+          <span className={`label-text self-start mb-3 ${className?.label_span}`}>{_label}</span>
           {Input}
         </label>
       </>
